Replace .web by .native
