--- src/components/SplashScreenHider/index.web.js
@@ -1,13 +1,0 @@
-import {useEffect} from 'react';
-import BootSplash from '../../libs/BootSplash';
-
-const SplashScreenHider = () => {
-    useEffect(() => {
-        BootSplash.hide();
-    }, []);
-
-    return null;
-};
-
-SplashScreenHider.displayName = 'SplashScreenHider';
-export default SplashScreenHider;
